php53: don't use same path for scandir and inidir, causes issues
Using the same dir causes the main php.ini to be loaded twice, resulting at
least in warnings about attempting to load the extension modules twice
(didn't do much further testing as this issue is big enough of itself but
wouldn't be surprised if more issues were caused).

Using /etc/php53/php as the scandir for now but I don't like it much, will be
looking for a better solution later on.